git user data from google

+++ server/controller/index.js
@@ -1,6 +1,17 @@
-const {register,getAllUsers,login,logout,deleteAllUsers}=require('./user');
+const {register,getAllUsers,login,logout,deleteAllUsers,googleAuth}=require('./user');
 const {getAllBlogs,addNewBlog,deleteAllBlogs}=require('./blog/blogs')
 const {client,server}=require('./errors')
 
-
-module.exports={register,getAllUsers,login,logout,getAllBlogs,addNewBlog,deleteAllBlogs,deleteAllUsers,client,server}
+module.exports={
+    register,
+    getAllUsers,
+    login,
+    logout,
+    getAllBlogs,
+    addNewBlog,
+    deleteAllBlogs,
+    deleteAllUsers,
+    googleAuth,
+    client,
+    server
+}
